% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pogodatar.R
\docType{data}
\name{types}
\alias{types}
\title{Pokémon types data.}
\format{Data frame with the following columns.  Many more could be added, feel free to suggest.
\describe{
 \item{types}{Types}
 \item{color}{Color conventionally used for the type, in hex format}
 }}
\source{
Colors from \url{https://bulbapedia.bulbagarden.net/wiki/Category:Type_color_templates}
}
\usage{
types
}
\description{
Pokémon types data.
}
\keyword{datasets}
